refactor: utils->extensions. Resource to foundations
