+++ app/src/main/java/com/aacademy/homework/ui/moviedetail/CastAdapter.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.aacademy.homework.data.model.Actor
 import com.aacademy.homework.databinding.LayoutCastItemBinding
+import com.aacademy.homework.extensions.loadImage
 import com.aacademy.homework.ui.moviedetail.CastAdapter.CastViewHolder
-import com.aacademy.homework.utils.extensions.loadImage
 import com.bumptech.glide.RequestManager
 
 class CastAdapter(val glide: RequestManager) : RecyclerView.Adapter<CastViewHolder>() {
